backend/requirements.txt
# DeepRAG — Knowledge Base + RAG Backend
# ============================================

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.0

# Vector store
chromadb>=0.4.22

# Embeddings & reranking
sentence-transformers>=2.2.2

# Document parsing (Docling)
docling>=2.0.0

# PDF rendering for scanned PDF detection + OCR pre-processing
pymupdf>=1.24.0

# Text processing
langchain-text-splitters>=0.0.1

# LLM providers
google-genai>=1.0.0
ollama>=0.4.0
openai>=1.0.0

# Knowledge Graph (LightRAG)
lightrag-hku>=1.0.0
neo4j>=5.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
aiofiles>=23.2.1
Pillow>=10.0.0

# Message queue
aio-pika>=9.4.0

# Object storage (MinIO / S3-compatible)
aioboto3>=13.0

# Digital signature extraction (PDF)
asn1crypto>=1.5.0

vllm